Rename restRequest class to RestRequest
Belongs to #5 and #18.
